refactor: split file utilities by responsibility (#4900)
