+++ server/lib/index.js
@@ -187,8 +187,14 @@ export * from './createKeyCachedQueue.js';
 export * from './createNewestWinsGuard.js';
 export * from './dataRoot.js';
 export * from './agentInstructionsFile.js';
-export * from './fileUtils.js';
+export * from './fileCore.js';
+export * as fileUtils from './fileUtils.js';
 export * from './fileWriteQueue.js';
+export * from './jsonIo.js';
+export * from './mimeTypes.js';
+export * from './paths.js';
+export * from './pathSafety.js';
+export * from './uploads.js';
 export * from './icloudFile.js';
 export * from './spawnCwd.js';
 export * from './schemaVersions.js';
